Dump the vue app object and imports into the new app file
This is as-is, with import paths updated.
